add blueprint users
