+++ schedule/__init__.py
@@ -2,9 +2,10 @@ from flask import Flask, render_template, flash, redirect, url_for
 from flask.views import MethodView
 from flask_migrate import Migrate
 from schedule.model import db
-from schedule.model import Person, Role, Timeinterval
+from schedule.model import Role, Timeinterval
+from schedule.user.models import Person
+from schedule.user.views import blueprint as user_blueprint
 
-from schedule.forms import LoginForm
 from flask_login import LoginManager, login_user, logout_user, current_user
 
 
@@ -17,7 +18,9 @@ def create_app():
 
     login_manager = LoginManager()
     login_manager.init_app(app)
-    login_manager.login_view = 'login'
+    login_manager.login_view = 'user.login'
+
+    app.register_blueprint(user_blueprint)
 
     @login_manager.user_loader
     def load_user(user_id):
@@ -39,38 +42,7 @@ def create_app():
             return render_template('smeny.html', title = title)
         else:
             flash('Log in for access', 'alert-info')
-            return redirect(url_for('login'))
-
-    @app.route('/login')
-    def login():
-        if current_user.is_authenticated:
-            return redirect(url_for('index'))
-        title = "Авторизация"
-        login_form = LoginForm()
-        return render_template('login.html', title = title, form=login_form)
-
-    @app.route('/logout')
-    def logout():
-        if current_user.is_authenticated:
-            logout_user()
-            flash('You are logged out', 'alert-secondary')
-        else:
-            flash('You are not logged in', 'alert-secondary')
-        return redirect(url_for('login'))
-
-    @app.route('/process-login', methods=['POST'])
-    def process_login():
-        form = LoginForm()
-
-        if form.validate_on_submit():
-            user = Person.query.filter(Person.username == form.login_name.data).first()
-            if user and user.check_password(form.password.data):
-                login_user(user)
-                flash('You are logged in', 'alert-success')
-                return redirect(url_for('index'))
-
-        flash('Username or Password are invalid', 'alert-danger')
-        return redirect(url_for('login'))
+            return redirect(url_for('user.login'))
 
     return app
