+++ schedule/templates/base.html
@@ -27,12 +27,12 @@
             </li>
             {% if not current_user.is_authenticated %}
             <li class="nav-item">
-              <a class="nav-item nav-link" href="{{ url_for('login') }}">Log in</a>
+              <a class="nav-item nav-link" href="{{ url_for('user.login') }}">Log in</a>
             </li>
             {% else %}
             <li class="nav-item">
               Logged in as {{ current_user.username }}
-              <a class="nav-item nav-link" href="{{ url_for('logout') }}">Log out</a>
+              <a class="nav-item nav-link" href="{{ url_for('user.logout') }}">Log out</a>
             </li>
             {% endif %}
           </div>
